add presentation module
