+++ app/src/androidTest/java/com/gomaa/nytimesarticles/ui/components/CoinLoaderTest.kt
@@ -3,8 +3,8 @@ package com.gomaa.nytimesarticles.ui.components
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import com.gomaa.nyarticles.presentation.ui.components.CoinLoader
-import com.gomaa.nyarticles.presentation.ui.components.CoinLoaderConstants.COIN_LOADER_TEST_TAG
+import com.gomaa.presentation.ui.components.CoinLoader
+import com.gomaa.presentation.ui.components.CoinLoaderConstants.COIN_LOADER_TEST_TAG
 import org.junit.Rule
 import org.junit.Test
 
@@ -15,8 +15,7 @@ class CoinLoaderTest {
     @Test
     fun verifyCoinLoaderIsDisplayed() {
         composeTestRule.setContent {
-            CoinLoader()
-
+           CoinLoader()
         }
         composeTestRule.onNodeWithTag(
             COIN_LOADER_TEST_TAG,
