+++ app/src/androidTest/java/com/gomaa/nytimesarticles/ui/components/EmptyStateUiTest.kt
@@ -3,8 +3,8 @@ package com.gomaa.nytimesarticles.ui.components
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import com.gomaa.nyarticles.presentation.ui.components.EmptyStateUi
-import com.gomaa.nyarticles.presentation.ui.components.EmptyStateUiConstants.EMPTY_STATE_TEST_TAG
+import com.gomaa.presentation.ui.components.EmptyStateUi
+import com.gomaa.presentation.ui.components.EmptyStateUiConstants.EMPTY_STATE_TEST_TAG
 import org.junit.Rule
 import org.junit.Test
 
@@ -15,8 +15,7 @@ class EmptyStateUiTest {
     @Test
     fun verifyEmptyStateIsDisplayed() {
         composeTestRule.setContent {
-            EmptyStateUi()
-
+           EmptyStateUi()
         }
         composeTestRule.onNodeWithTag(
             EMPTY_STATE_TEST_TAG,
