+++ app/src/androidTest/java/com/gomaa/nytimesarticles/ui/components/ErrorStateUiTest.kt
@@ -3,8 +3,8 @@ package com.gomaa.nytimesarticles.ui.components
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import com.gomaa.nyarticles.presentation.ui.components.ErrorStateUi
-import com.gomaa.nyarticles.presentation.ui.components.ErrorStateUiConstants.ERROR_STATE_TEST_TAG
+import com.gomaa.presentation.ui.components.ErrorStateUi
+import com.gomaa.presentation.ui.components.ErrorStateUiConstants.ERROR_STATE_TEST_TAG
 import org.junit.Rule
 import org.junit.Test
 
@@ -16,7 +16,6 @@ class ErrorStateUiTest {
     fun verifyErrorStateIsDisplayed() {
         composeTestRule.setContent {
             ErrorStateUi()
-
         }
         composeTestRule.onNodeWithTag(
             ERROR_STATE_TEST_TAG,
