+++ app/src/androidTest/java/com/gomaa/nytimesarticles/ui/components/FilterChipRowTest.kt
@@ -3,8 +3,8 @@ package com.gomaa.nytimesarticles.ui.components
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import com.gomaa.nyarticles.presentation.ui.components.FilterChipRow
-import com.gomaa.nyarticles.presentation.ui.components.FilterComponentConstants
+import com.gomaa.presentation.ui.components.FilterChipRow
+import com.gomaa.presentation.ui.components.FilterComponentConstants
 import com.gomaa.utils.initializeFilterOptions
 import org.junit.Rule
 import org.junit.Test
@@ -16,7 +16,9 @@ class FilterChipRowTest {
     @Test
     fun verifyFilterChipsIsDisplayed() {
         composeTestRule.setContent {
-            FilterChipRow(filterOptions = initializeFilterOptions(), {})
+            FilterChipRow(
+                filterOptions = initializeFilterOptions(),
+                {})
 
         }
         composeTestRule.onNodeWithTag(
