+++ features/nyarticles/presentation/build.gradle.kts
@@ -1,33 +1,25 @@
 plugins {
-    id("com.android.application")
+    id("com.android.library")
     id("org.jetbrains.kotlin.android")
     id("kotlin-kapt")
     kotlin("kapt")
+    id("com.google.devtools.ksp")
     id("com.google.dagger.hilt.android")
+    id("org.jetbrains.kotlin.plugin.serialization") version "1.8.22"
 }
 
 android {
-    namespace = "com.gomaa.nytimesarticles"
+    namespace = "com.gomaa.presentation"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.gomaa.nytimesarticles"
         minSdk = 26
-        targetSdk = 34
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        vectorDrawables {
-            useSupportLibrary = true
-        }
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     buildTypes {
-        debug {
-            isMinifyEnabled = false
-            isDebuggable = true
-        }
         release {
             isMinifyEnabled = false
             proguardFiles(
@@ -49,38 +41,31 @@ android {
     composeOptions {
         kotlinCompilerExtensionVersion = "1.5.1"
     }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-    configurations.all {
-        resolutionStrategy {
-            force("androidx.test.ext:junit:1.1.3")
-            force("androidx.test.espresso:espresso-core:3.4.0")
-        }
-    }
 }
 
 dependencies {
+    implementation(project(":core"))
     implementation(project(":design-system"))
-    implementation(project(":features:nyarticles:presentation"))
     implementation(project(":utils"))
+    implementation(project(":features:nyarticles:data"))
+    implementation(project(":features:nyarticles:domain"))
     implementation("androidx.core:core-ktx:1.13.1")
-    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.8.0")
-    implementation("androidx.activity:activity-compose:1.9.0")
-    implementation(platform("androidx.compose:compose-bom:2023.08.00"))
-    implementation("androidx.compose.ui:ui")
-    implementation("androidx.compose.ui:ui-graphics")
-    implementation("androidx.compose.ui:ui-tooling-preview")
+    implementation("androidx.appcompat:appcompat:1.6.1")
+    implementation("com.google.android.material:material:1.12.0")
+    testImplementation("junit:junit:4.13.2")
+    androidTestImplementation("androidx.test.ext:junit:1.1.5")
+    androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
     implementation("androidx.compose.material3:material3")
-
+    // Coroutine
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.8.1")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.8.1")
     implementation("com.google.dagger:hilt-android:2.49")
     implementation("androidx.compose.ui:ui-test-junit4-android:1.6.7")
     testImplementation("org.junit.jupiter:junit-jupiter:5.8.2")
     kapt("com.google.dagger:hilt-compiler:2.48.1")
     implementation("androidx.hilt:hilt-navigation-compose:1.2.0")
 
+    implementation("io.coil-kt:coil-compose:2.6.0")
     implementation("org.mockito.kotlin:mockito-kotlin:4.0.0")
     implementation("org.mockito:mockito-core:4.0.0")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-test:1.8.1")
