+++ features/nyarticles/presentation/src/main/java/com/gomaa/presentation/model/ArticleUiEntity.kt
@@ -1,0 +1,12 @@
+package com.gomaa.presentation.model
+
+data class ArticleUiEntity(
+    val title: String,
+    val publishedDate: String,
+    val byLine: String,
+    val url: String,
+    val smallThumbnail:String,
+    val mediumThumbnail:String,
+    val largeThumbnail:String,
+    val publishedAgo:String
+)
