+++ features/nyarticles/presentation/src/main/java/com/gomaa/presentation/viewmodel/PopularArticlesEvent.kt
@@ -1,0 +1,7 @@
+package com.gomaa.presentation.viewmodel
+
+import com.gomaa.core.presentation.UiEvent
+
+sealed class PopularArticlesEvent : UiEvent {
+    data class FetchPopularArticles(val period: Int) : PopularArticlesEvent()
+}
