+++ features/nyarticles/presentation/src/main/res/drawable/close_circle.xml
