+++ features/nyarticles/presentation/src/main/res/drawable/error.xml
