+++ features/nyarticles/presentation/src/main/res/drawable/illustration_empty.xml
