+++ features/nyarticles/presentation/src/main/res/drawable/open_in_new.xml
