--- features/nyarticles/src/main/java/com/gomaa/nyarticles/presentation/viewmodel/PopularArticlesEvent.kt
@@ -1,7 +1,0 @@
-package com.gomaa.nyarticles.presentation.viewmodel
-
-import com.gomaa.core.presentation.UiEvent
-
-sealed class PopularArticlesEvent : UiEvent {
-    data class FetchPopularArticles(val period: Int) : PopularArticlesEvent()
-}
