--- features/nyarticles/src/main/res/drawable/close_circle.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-    <path
-        android:pathData="M3.75,12C3.75,7.443 7.443,3.75 12,3.75C16.557,3.75 20.25,7.443 20.25,12C20.25,16.557 16.557,20.25 12,20.25C7.443,20.25 3.75,16.557 3.75,12ZM12,2.25C6.615,2.25 2.25,6.615 2.25,12C2.25,17.385 6.615,21.75 12,21.75C17.385,21.75 21.75,17.385 21.75,12C21.75,6.615 17.385,2.25 12,2.25ZM15.36,8.64C15.653,8.933 15.653,9.407 15.36,9.7L13.061,12L15.36,14.3C15.653,14.592 15.653,15.067 15.36,15.36C15.067,15.653 14.592,15.653 14.3,15.36L12,13.061L9.7,15.36C9.407,15.653 8.932,15.653 8.64,15.36C8.347,15.067 8.347,14.592 8.64,14.3L10.939,12L8.64,9.7C8.347,9.407 8.347,8.933 8.64,8.64C8.932,8.347 9.407,8.347 9.7,8.64L12,10.939L14.3,8.64C14.592,8.347 15.067,8.347 15.36,8.64Z"
-        android:fillColor="#000000"
-        android:fillType="evenOdd"/>
-</vector>
