--- features/nyarticles/src/main/res/drawable/error.xml
@@ -1,5 +1,0 @@
-<vector android:height="64dp" android:viewportHeight="32"
-    android:viewportWidth="32" android:width="64dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#D72828" android:pathData="M16,16m-16,0a16,16 0,1 1,32 0a16,16 0,1 1,-32 0"/>
-    <path android:fillColor="#E6E6E6" android:pathData="M14.5,25h3v-3h-3V25zM14.5,6v13h3V6H14.5z"/>
-</vector>
