--- features/nyarticles/src/main/res/drawable/illustration_empty.xml
@@ -1,32 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="96dp"
-    android:height="96dp"
-    android:viewportWidth="96"
-    android:viewportHeight="96">
-    <path
-        android:pathData="M20,86a28,6 0,1 0,56 0a28,6 0,1 0,-56 0z"
-        android:strokeAlpha="0.3"
-        android:fillColor="#BD80FA"
-        android:fillAlpha="0.3"/>
-    <path
-        android:pathData="M38.18,61.246C49.073,68.629 64.015,67.495 73.665,57.845C84.599,46.91 84.599,29.182 73.665,18.247C62.73,7.312 45.001,7.312 34.067,18.247C24.761,27.552 23.375,41.778 29.907,52.549L16.243,66.213C13.9,68.556 13.9,72.355 16.243,74.699C18.586,77.042 22.385,77.042 24.728,74.699L38.18,61.246Z"
-        android:fillColor="#D1A7FB"
-        android:fillType="evenOdd"/>
-    <path
-        android:pathData="M73.799,56.799C64.31,66.288 49.705,67.543 38.863,60.563L32.43,66.996C32.632,62.356 31.436,57.667 28.843,53.613L30.391,52.065C23.463,41.23 24.733,26.669 34.201,17.201C45.136,6.266 62.864,6.266 73.799,17.201C84.734,28.136 84.734,45.864 73.799,56.799Z"
-        android:fillColor="#A358EE"
-        android:fillType="evenOdd"/>
-    <path
-        android:pathData="M73.665,56.845L73.134,56.315L73.134,56.315L73.665,56.845ZM38.783,60.644L39.188,60.012C38.891,59.822 38.502,59.864 38.253,60.113L38.783,60.644ZM73.665,17.247L73.134,17.777L73.134,17.777L73.665,17.247ZM34.067,17.247L33.536,16.717L33.536,16.717L34.067,17.247ZM30.291,52.165L30.822,52.695C31.071,52.445 31.113,52.056 30.922,51.759L30.291,52.165ZM16.243,66.213L16.773,66.743L16.773,66.743L16.243,66.213ZM16.243,74.699L16.773,74.168L16.243,74.699ZM24.728,74.699L25.258,75.229L25.258,75.229L24.728,74.699ZM73.134,56.315C63.916,65.533 49.733,66.767 39.188,60.012L38.379,61.275C49.505,68.402 64.467,67.103 74.195,57.375L73.134,56.315ZM73.134,17.777C83.776,28.419 83.776,45.673 73.134,56.315L74.195,57.375C85.423,46.148 85.423,27.944 74.195,16.717L73.134,17.777ZM34.597,17.777C45.239,7.136 62.493,7.136 73.134,17.777L74.195,16.717C62.967,5.489 44.764,5.489 33.536,16.717L34.597,17.777ZM30.922,51.759C24.142,41.21 25.368,27.006 34.597,17.777L33.536,16.717C23.798,26.455 22.507,41.44 29.66,52.57L30.922,51.759ZM16.773,66.743L30.822,52.695L29.761,51.634L15.712,65.683L16.773,66.743ZM16.773,74.168C14.723,72.118 14.723,68.794 16.773,66.743L15.712,65.683C13.076,68.319 13.076,72.593 15.712,75.229L16.773,74.168ZM24.198,74.168C22.147,76.218 18.823,76.218 16.773,74.168L15.712,75.229C18.348,77.865 22.622,77.865 25.258,75.229L24.198,74.168ZM38.253,60.113L24.198,74.168L25.258,75.229L39.313,61.174L38.253,60.113Z"
-        android:fillColor="#181877"/>
-    <path
-        android:pathData="M54,37m-23,0a23,23 0,1 1,46 0a23,23 0,1 1,-46 0"
-        android:fillColor="#E8D3FD"/>
-    <path
-        android:pathData="M73.91,25.478C69.692,20.563 63.435,17.45 56.45,17.45C43.747,17.45 33.45,27.747 33.45,40.45C33.45,44.648 34.575,48.584 36.54,51.972C33.086,47.948 31,42.718 31,37C31,24.298 41.298,14 54,14C62.504,14 69.931,18.616 73.91,25.478Z"
-        android:fillColor="#D1A7FB"
-        android:fillType="evenOdd"/>
-    <path
-        android:pathData="M63,31m-7,0a7,7 0,1 1,14 0a7,7 0,1 1,-14 0"
-        android:fillColor="#F8F1FE"/>
-</vector>
